partition

Module multi_dim

Structs

MultiDimPartitionRule

Multi-Dimiension partition rule. RFC here

RuleChecker 🔒

Check if the rule set covers all the possible values.

SplitPoint 🔒

Helper for RuleChecker

Constants

DEFAULT_REGION 🔒

The default region number when no partition exprs are matched.

Type Aliases

Axis 🔒

Helper for RuleChecker

PhysicalExprCache 🔒