Trait partition::partition::PartitionRule

pub trait PartitionRule: Sync + Send {
    // Required methods
    fn as_any(&self) -> &dyn Any;
    fn partition_columns(&self) -> Vec<String>;
    fn find_region(&self, values: &[Value]) -> Result<RegionNumber>;
    fn find_regions_by_exprs(
        &self,
        exprs: &[PartitionExpr]
    ) -> Result<Vec<RegionNumber>>;
}

Required Methods

fn as_any(&self) -> &dyn Any

fn partition_columns(&self) -> Vec<String>

fn find_region(&self, values: &[Value]) -> Result<RegionNumber>

Finds the target region by the partition values.

Note that the values should have the same length as the partition_columns.

fn find_regions_by_exprs( &self, exprs: &[PartitionExpr] ) -> Result<Vec<RegionNumber>>

Finds the target regions by the partition expressions.

Note that the exprs should have the same length as the partition_columns.